[refactor] vo 패키지명 변경
* 기존 car -> vo 패키지로 이름 변경
